оптимизация памяти
сделал 2 клетки лабиринта на байт, убрал указатели много где, и всё равно занимает аж 60% памяти.

+++ ArduinoCode/include/Encoder.h
@@ -4,9 +4,9 @@
 
 struct EncoderParameters
 {
-    int CLOCK_A_PIN;
-    int B_PIN;
-    int DIR;
+    int8_t CLOCK_A_PIN;
+    int8_t B_PIN;
+    int8_t DIR;
     void (*ISRfunc)();
 };
